Add a tool to check i18n files
check_i18n_files.html is a page which compares the given i18n file to the
reference template (the english version).
Each language has to be check manually.

It provides key words that are missing in the file and key words that have
not been translated (empty string or other null value).